src/app/components/add-todo/add-todo.component.ts
import { Component } from '@angular/core';
import { FormBuilder, FormGroup, Validators } from '@angular/forms';
import { TodoService } from '../../services/todo.service';

@Component({
  selector: 'app-add-todo',
  templateUrl: './add-todo.component.html',
  styleUrl: './add-todo.component.css'
})
export class AddTodoComponent {
  todoForm: FormGroup;

  constructor(private fb: FormBuilder, private todoService: TodoService) {
    this.todoForm = this.fb.group({
      title: ['', Validators.required]
    });
  }

  addTodo(): void {
    if (this.todoForm.valid) {
      this.todoService.addTodo(this.todoForm.value.title);
      this.todoForm.reset();
    }
  }

}
